requirements.txt
tensorflow-gpu
tensorflow_hub
numpy
pandas
tokenization
sentencepiece
sklearn